+++ packages/integration-sdk-runtime/src/__tests__/index.test.ts
@@ -1,0 +1,11 @@
+import { BucketMap, FileSystemGraphObjectStore } from '../index';
+
+describe('#storage', () => {
+  test('should expose BucketMap', () => {
+    expect(BucketMap).not.toEqual(undefined);
+  });
+
+  test('should expose FileSystemGraphObjectStore', () => {
+    expect(FileSystemGraphObjectStore).not.toEqual(undefined);
+  });
+});