+++ packages/integration-sdk-runtime/src/storage/FileSystemGraphObjectStore/index.ts
@@ -1,1 +1,2 @@
 export * from './FileSystemGraphObjectStore';
+export * from './BucketMap';